% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_utils.R
\name{as_biallelic}
\alias{as_biallelic}
\title{Biallelic SNPs.}
\usage{
as_biallelic(snps, toupper = FALSE)
}
\arguments{
\item{snps}{A list of locus-specific SNP matrices, as produced by \link{get_snps}.}

\item{toupper}{Logical, whether to enforce representation in upper case letters.}
}
\value{
A list of modified locus-specific SNP matrices.
}
\description{
Makes SNPs biallelic by marking the third and fourth alleles as missing data.
}
