% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascertbias.R
\name{ascertbias}
\alias{ascertbias}
\title{Ascertainment bias.}
\usage{
ascertbias(loci, info, allele = "_[012]$", thin = 1)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{matrix} or \code{DNAbin}.}

\item{info}{a data frame or matrix listing individual labels (1st column) and group labels (2nd column).}

\item{allele}{regular expression, allele identifier in rownames of locus-specific alignments.}

\item{thin}{a proportion of loci to be retained or a subsampling frequency (if \code{thin > 1}).}
}
\value{
A list with components \code{occupancy} (proportion of successfully sequenced individuals),
\code{mdist} (mean uncorrected genetic distance), \code{nalleles} (the number of distinct haplotypes),
\code{nsnps} (the number of SNPs), \code{npis} (the number of parsimony-informative SNPs),
\code{heterozygosity} (mean heterozygosity), \code{spoccupancy} (proportion of species or other taxa represented
at the locus), \code{mintra} (mean intra-specific uncorrected genetic distance) and \code{thin} (the \code{thin} argument).
The components are numeric vectors of the same length as the number of loci after thinning.
}
\description{
Calculates statistics helping to detect of ascertainment bias due to different occupancy across loci.
}
