% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_utils.R
\name{bpcoverage}
\alias{bpcoverage}
\title{Base-pair occupancy.}
\usage{
bpcoverage(locus)
}
\arguments{
\item{locus}{a locus-specific alignment.}
}
\value{
A numeric vector with proportions of non-missing data along the locus.
}
\description{
For every base pair it calculates its occupancy, i.e., proportion of sequences with non-missing data.
}
