% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_snps.R
\name{center_binary_snps}
\alias{center_binary_snps}
\title{Center SNP genotypes.}
\usage{
center_binary_snps(b, counts = TRUE)
}
\arguments{
\item{b}{matrix, the output of \link{make_binary_snps}.}

\item{counts}{logical, whether individual genotype is represented in a single row by counts of '1' alleles.}
}
\value{
A matrix with centered binary coded genotypes.
}
\description{
Centers binary coded SNP genotypes, so in the 'counts' representation, heterozygote takes value of 0 and homozygotes -1 and 1.
}
