% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{classifyseq}
\alias{classifyseq}
\title{Classify sequences.}
\usage{
classifyseq(info, haps, col.names = c("seq", "tax"), na.rm = FALSE)
}
\arguments{
\item{info}{a data frame with information about classification of haplotypes to taxa (species, clades, lineages)
or a name of file with such classification. Haplotype names are in the 1st column, taxa in the 2nd column.}

\item{haps}{either a data frame with assignment of sequences to haplotypes (as produced by \link{haplotypes})
or a name of file with such an assignment. Haplotype names are in the 1st column, sequence names in the 2nd column.}

\item{col.names}{column names of the output data frame.}

\item{na.rm}{logical, whether to omit sequences whose classification to taxa cannot be established
from the classification of haplotypes.}
}
\value{
A data frame with similar to \code{info} input, but having in the 1st column sequence names
and in the 2nd column assignment to taxa, based on the correspondence of haplotype names.
}
\description{
Classify sequences based on their identity to already classified ones.
}
