% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_utils.R
\name{clean_data}
\alias{clean_data}
\title{Cleaning of data.}
\usage{
clean_data(loci, remove = "empty", numeric = FALSE)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{remove}{\code{"empty"} (default) means removing all empty rows and columns, which are completely empty,
whereas \code{"rows"} and \code{"columns"} means removing whole rows or columns with any missing data.}

\item{numeric}{logical, whether are the data numeric (e.g. binary representation of biallelic SNPs).}
}
\value{
A numeric vector of locus-specific occupancies.
}
\description{
Removes missing data for (an) alignment(s).
}
