% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_diploid_genotypes.R
\name{collapse_diploid_genotypes}
\alias{collapse_diploid_genotypes}
\title{Collapsed diploid genotypes.}
\usage{
collapse_diploid_genotypes(loci, allele = "_[012]$")
}
\arguments{
\item{loci}{an alignment in \code{matrix} or \code{DNAbin} format or a list of them.}

\item{allele}{regular expression specifying allele indicator.}
}
\value{
A modified alignment in \code{matrix} or \code{DNAbin} format or a list of them.
}
\description{
Collapses diploid genotypes into a single sequence using ambiguity codes.
}
