% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_utils.R
\name{count_snps}
\alias{count_snps}
\title{Counting SNPs.}
\usage{
count_snps(indices, pis = FALSE, nalleles = 4)
}
\arguments{
\item{indices}{A list of SNP indices produced by \link{find_snps}.}

\item{pis}{Logical, whether to retain only parsimony-informative sites.}

\item{nalleles}{Numeric, the maximum no. of alleles per variable site.}
}
\value{
A list of locus-specific SNP matrices, but with just one site (column) per locus. Invariant loci a discarded.
}
\description{
A utility function counting SNPs specified the function arguments.
}
