% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_utils.R
\name{coverage}
\alias{coverage}
\title{Coverage.}
\usage{
coverage(loci, allele = "_[012]$", indiv = NULL, prop = TRUE)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{allele}{a regular expression defining an allele identifier (allows to distinguish individual IDs).}

\item{indiv}{an optional character vector giving complete list of individual IDs (useful to get also
individuals with zero coverage represented in the output).}

\item{prop}{logical, whether to express the coverages as proportions (default) or counts}
}
\value{
A numeric vector of individual-specific coverages.
}
\description{
Calculates coverage, i.e., proportion of loci for which an individual was successfully sequenced.
}
