% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_traces.R
\name{discard_burnin}
\alias{discard_burnin}
\title{Discarding burn-in.}
\usage{
discard_burnin(x, relburnin, absburnin)
}
\arguments{
\item{x}{an object of class \code{multiPhylo} or a data frame.}

\item{relburnin}{burn-in period specified as a proportion of posterior sample.}

\item{absburnin}{burn-in period specified as the number of sampled values, overrides \code{relburnin}.}
}
\value{
A modified input, with the burn-in part discarded.
}
\description{
Discards an initial, burn-in, part of the posterior sample.
}
