% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_snps.R
\name{dissim.snps}
\alias{dissim.snps}
\title{Dissimilarity of SNP genotypes.}
\usage{
dissim.snps(b, allele = "_[[:digit:]]$")
}
\arguments{
\item{b}{matrix, the output of \link{make_binary_snps}.}

\item{allele}{regular expression, allele identifier in the rownames of snps.}
}
\value{
A matrix of pairwise dissimilarities.
}
\description{
Calculates matrix of pairwise genotype dissimilarities. The dissimilarity is just number of different alleles,
averaged over alleles and omitting data missing in either of the two compared genotypes.
}
