% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergentseq.R
\name{divergentseq}
\alias{divergentseq}
\title{Most divergent sequences.}
\usage{
divergentseq(
  input,
  n,
  minlen = 0,
  fasta = NULL,
  return = TRUE,
  outgroup = NULL,
  reserved = NULL,
  model = "raw"
)
}
\arguments{
\item{input}{a character string with name of the file in format accepted by \link{read_seq}
or a sequence alignment in \code{DNAbin} or \code{matrix} format or an output of \link{haplotypes}.}

\item{n}{number of sequences to retain.}

\item{minlen}{numeric the minimum required length of sequence.}

\item{fasta}{character, a name of the output .fasta file with the alignment of haplotypes.}

\item{return}{logical, whether to return the output as a list.}

\item{outgroup}{a character vector with outgroup sequence names or their unique identifier(s),
e.g., a character string \code{"outgroup"}.}

\item{reserved}{a character vector listing sequences that must be included in the selection.}

\item{model}{a character string specifying the nucleotide substitution model used to correct for multiple
mutations when calculating pairwise distences between sequences; deafult is \code{"raw"} (no correction).}
}
\value{
An alignment of \code{n} retained sequences (in \code{DNAbin} or \code{matrix} format, depending on the input).
}
\description{
Selection of a specified number of sequences, which are most divergent from each other.
}
\details{
The algorithm proceeds by stepwise elimination of one of the most similar sequences in the alignment.
}
