% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow.R
\name{elbow}
\alias{elbow}
\title{Elbow method.}
\usage{
elbow(y, x = seq_along(y))
}
\arguments{
\item{y}{numeric, eigenvalues from PCA.}

\item{x}{ranks of the principal components (defaults to \code{seq_along(y)}).}
}
\value{
A list with components \code{x} and \code{y} (the input data), \code{b} (breakpoints), \code{rmse} (root mean squared error)
and \code{best} (the optimal breakpoint, corresponding to \code{b[which.min(rmse)]}). It is assigned a class \code{elbow}.
}
\description{
An elbow method to determine the number of important principal components.
}
\details{
The method finds a breakpoint separating first principal components accounting for most of the variation
from the remaining ones. The function is an implementation of approach described by Salvador and Chan 2004
(https://doi.org/10.1109/ICTAI.2004.50). It can be used to find breakpoints also in other analogous data.
}
