% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyseq.R
\name{epaformat}
\alias{epaformat}
\title{EPA format.}
\usage{
epaformat(
  taxa,
  haps,
  epa,
  file,
  return = TRUE,
  na.rm = FALSE,
  retain = TRUE,
  method
)
}
\arguments{
\item{taxa}{a data frame with information about classification of haplotypes to taxa (species, clades, lineages)
or a name of file with such classification. Haplotype names are in the 1st column, taxa in the 2nd column.}

\item{haps}{either an output of \link{haplotypes} or a data frame with assignment of sequences to haplotypes
(\code{assign} component of the list produced by \link{haplotypes} or a name of tab-delimited file with such data).
Sequence names are in the 1st column, haplotype names in the 2nd column.}

\item{epa}{an output of \code{\link[epaclades:classify_jplace]{epaclades::classify_jplace()}}, \code{\link[epaclades:classify_sequences]{epaclades::classify_sequences()}} or
\code{\link[epaclades:best_placement]{epaclades::best_placement()}} or a name of tab-delimited file with output of \code{\link[epaclades:best_placement]{epaclades::best_placement()}}.
If missing, the epa formatting is still imposed to the output based on \code{taxa} and \code{haps}.}

\item{file}{character string, the name of the output file.}

\item{return}{logical, whether to return the resulting sequence alignment.}

\item{na.rm}{logical, whether to omit sequences that could not be classified into taxa.}

\item{retain}{logical, whether to retain all columns of \code{taxa} data frame or just the first two of them.}

\item{method}{optional character string, method of classification into units predefined in \code{taxa}.}
}
\value{
A data frame with six columns labelled "id" (sequence name), "clade" (taxon), "method",
"position" (stem or crown), "probability" (for the position), "totalprob" (summed over stem and crown branches).
The last three columns are relevant only for classification by phylogenetic placement.
The "method" column takes three possible values: "predefined" (or the name of the original delimitation method
from \code{method} argument), "epa" (classification by phylogenetic placement) or "match" (if the classification
was inherited from another through sequence matching, possibly with the associated probabilities).
}
\description{
Summarizes classification of sequences into taxa in a format similar to output of \code{\link[epaclades:best_placement]{epaclades::best_placement()}}.
}
