% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyseq.R
\name{epainput}
\alias{epainput}
\title{EPA input.}
\usage{
epainput(ref, query, seq, file, return = FALSE)
}
\arguments{
\item{ref}{a character vector with IDs of reference sequences or a single tree (readable by \link{read_tree}),
whose tip labels are used as IDs of reference sequences.}

\item{query}{a character vector with IDs of query sequences. Also it could be an object accepted as \code{haps} argument
by \link{findqueryseq}, i.e., either an output of \link{haplotypes} or a data frame with assignment of sequences to haplotypes
(\code{assign} component of the list produced by \link{haplotypes} or a name of tab-delimited file with such data).}

\item{seq}{either a a sequence alignment or a name of .fasta file which contains both reference and query sequences.}

\item{file}{character string, the name of .fasta file with the resulting alignment.}

\item{return}{logical, whether to return the resulting sequence alignment.}
}
\value{
If \code{return=TRUE}, the resulting sequence alignment in the same format in which it was supplied or imported in.
}
\description{
Prepares sequence file to be used as an input for phylogenetic placement, i.e., including reference and query sequences.
}
