% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytree.R
\name{estimate_brlen}
\alias{estimate_brlen}
\title{Branch lengths.}
\usage{
estimate_brlen(
  phy,
  trees,
  strict = FALSE,
  ultrametric = FALSE,
  fun = "mean",
  thin = 1,
  p = 0.95,
  interval = "HPD"
)
}
\arguments{
\item{phy}{a \code{phylo} object with the summary tree (branch lengths are omitted, even if present).}

\item{trees}{a posterior sample of rooted trees in a \code{multiPhylo} object or a list of \code{phylo} objects
or an object of class \code{bpp} or a name of file with the posterior sample.}

\item{strict}{logical, whether to estimate branch lengths only from the trees, whose topology is strictly
identical with that of a summary tree.}

\item{ultrametric}{logical, whether the \code{trees} are ultrametric. If set to \code{FALSE} (default), it is actually decided
according to tip-to-root distances in the first of the trees.}

\item{fun}{character string, name of the function estimating branch lengths (default is \code{"mean"}).}

\item{thin}{a proportion of mcmc steps to be retained or a subsampling frequency (if \code{thin > 1}).}

\item{p}{a proportion of posterior density to be included in HPD/CPD interval of branch length estimate.
If \code{p=0}, no interval is estimated. The default is now 0, but in future it may change to 0.95.}

\item{interval}{a type of interval to be calculated, either \code{"HPD"} (highest posterior density, default)
or \code{"CPD"} (central posterior density).}
}
\value{
An object of class \code{phylo} amended by branch lengths estimated from the provided sample of trees.
}
\description{
Estimates branch lengths from a sample of trees.
}
