% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extraparams}
\alias{extraparams}
\title{Extra parameters of STRUCTURE}
\format{
A character vector with the lines of 'extraparams' template provided with STRUCTURE.
}
\source{
https://web.stanford.edu/group/pritchardlab/structure.html
}
\usage{
extraparams
}
\description{
Template for extra parameters file of STRUCTURE.
}
\keyword{datasets}
