% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_utils.R
\name{find_snps}
\alias{find_snps}
\title{Finding SNPs.}
\usage{
find_snps(loci, allele = NULL)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"},
possibly an output of \link{import_seq}.}

\item{allele}{a regular expression defining an allele identifier (allows to distinguish individual IDs).}
}
\value{
A list of lists with components \code{"snp"} (numeric, indices of SNP positions),
\code{"pis"} (logical, whether the SNPs are parsimony-informative)
and \code{"nalleles"} (the number of distinct haplotypes = alleles in the alignment).
}
\description{
Finds variable positions (a.k.a. single nucleotide polymorphisms or SNPs) in sequence alignements.
}
