% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_utils.R
\name{get_snps}
\alias{get_snps}
\title{Extracting SNPs.}
\usage{
get_snps(
  loci,
  indices = NULL,
  pis = FALSE,
  nalleles = 4,
  rm.invariant = TRUE,
  toupper = FALSE,
  subset = TRUE,
  allele = NULL
)
}
\arguments{
\item{loci}{A list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{indices}{A list of SNP indices produced by \link{find_snps}. If not provided, it is obtained internally.}

\item{pis}{Logical, whether to retain only parsimony-informative sites. If allele is specified,
it requires the minority variant to be present in at least two individuals.}

\item{nalleles}{Numeric, the maximum no. of alleles per variable site,
i.e., \code{nalleles=2} means extracting just biallelic SNPs.}

\item{rm.invariant}{Logical, whether to remove invariant loci from the list.}

\item{toupper}{Logical, whether to enforce representation in upper case letters.}

\item{subset}{A vector defining subset of loci; logical, numeric (indices of loci) or character (names of loci).}

\item{allele}{A regular expression defining an allele identifier (allows to distinguish individual IDs).}
}
\value{
A list of locus-specific SNP matrices (subsets of original alignments), each with attributes \code{"snp"},
\code{"pis"} and\code{"nalleles"} corresponding to \code{indices} argument.
}
\description{
Extracts specified SNPs from locus-specific alignments.
}
