% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_utils.R
\name{grcoverage}
\alias{grcoverage}
\title{Group coverage.}
\usage{
grcoverage(loci, info, allele = "_[012]$")
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{info}{a data frame or matrix listing individual labels (1st column) and group labels (2nd column).}

\item{allele}{a regular expression defining an allele identifier (allows to distinguish individual IDs).}
}
\value{
A numeric vector of group-specific coverages.
}
\description{
Calculates group coverage, i.e., proportion of loci for which at least one individual from given groups
(species, populations, lineages ...) was successfully sequenced.
}
