% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_utils.R
\name{groccupancy}
\alias{groccupancy}
\title{Group occupancy.}
\usage{
groccupancy(loci, info, allele = "_[012]$")
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{info}{a data frame or matrix listing individual labels (1st column) and group labels (2nd column).}

\item{allele}{regular expression, allele identifier in the rownames of snps.}
}
\value{
A numeric vector of locus-specific group occupancies.
}
\description{
Calculates group occupancy, i.e., proportion of groups (species, populations, lineages ...)
for which a locus was successfully sequenced from at least single individual.
}
