% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{haplotypes}
\alias{haplotypes}
\title{Unique haplotypes.}
\usage{
haplotypes(
  input,
  fasta = NULL,
  txt = NULL,
  return = TRUE,
  outgroup = NULL,
  minpres = 1,
  maxmiss = 1
)
}
\arguments{
\item{input}{a character string with file name (fasta or phylip interleaved format)
or sequence alignment in \code{DNAbin} or \code{matrix} format.}

\item{fasta}{character, a name of the output .fasta file with the alignment of haplotypes.}

\item{txt}{character, a name of the output .txt file with assignment of sequences to haplotypes.}

\item{return}{logical, whether to return the output as a list.}

\item{outgroup}{a character vector with outgroup sequence names or their unique identifier(s),
e.g., a character string \code{"outgroup"}. The outgroups are excluded from haplotype identification,
but presented in the output.}

\item{minpres}{numeric, a minimum required proportion of nucleotides non-missing in the sequence
under consideration and also in a haplotype it could be assigned to (default is 1).}

\item{maxmiss}{numeric, a maximum acceptable proportion of nucleotides missing in the sequence
under consideration, but non-missing in a haplotype it could be assigned to (default is 1).}
}
\value{
If \code{return == TRUE}, a list with components \code{haplotypes} (alignment of haplotypes),
\code{assign} (assignment of sequences to haplotypes), \code{no} (number of unique haplotypes found)
and \code{outgroup} (alignment of outgroup sequences).
}
\description{
Extracts unique haplotypes from an alignment.
}
\details{
Under the default values of \code{minpres} and \code{maxmiss} a shorter sequence is considered identical
to a longer one, if it does not contain any nucleotides missing in the longer sequence, but not vice versa.
If \code{minpres=1} and \code{maxmiss=0}, a precise match is required.
}
