% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heliophobius_cytb}
\alias{heliophobius_cytb}
\title{Cytochrome b sequences of Heliophobius}
\format{
A matrix of mode character with 28 rows (individuals) and 1140 columns (sites).
}
\source{
Uhrová et al. 2022 (Mol. Phyl Evol. 167: 107337).
}
\usage{
heliophobius_cytb
}
\description{
An alignment of 25 cytochrome b (CYTB) sequences of Silvery Mole-rat (Heliophobius)
and 3 outgroups (Georychus).
}
\keyword{datasets}
