% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heliophobius_info}
\alias{heliophobius_info}
\title{Information about CYTB of Heliophobius}
\format{
A data frame with 25 rows (individuals) and 6 columns.
}
\source{
Uhrová et al. 2022 (Mol. Phyl Evol. 167: 107337).
}
\usage{
heliophobius_info
}
\description{
Information about CYTB sequences of Heliophobius (\code{heliophobius_cytb} data set)
including individual ID, locality of origin and its geographical coordinates.
}
\keyword{datasets}
