% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_snps.R
\name{impute_binary_snps}
\alias{impute_binary_snps}
\title{Imputation of biallelic SNPs.}
\usage{
impute_binary_snps(b, d = NULL, k = 3)
}
\arguments{
\item{b}{matrix, the output of \link{make_binary_snps}.}

\item{d}{optional, matrix of pairwise genotype dissimilarities.}

\item{k}{numeric, the number of nearest neighbor genotypes used for imputation of missing data.}
}
\value{
A modified matrix with binary coded genotypes.
}
\description{
Imputes missing biallelic snps.
}
