% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_utils.R
\name{loclen}
\alias{loclen}
\title{Locus lengths.}
\usage{
loclen(loci, rm.gaps = FALSE, outgroup = NULL)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{rm.gaps}{logical, whether to exclude gaps from calculating of lengths.}

\item{outgroup}{identifier(s) of outgroup(s) not to be taken into account when detecting gaps.}
}
\value{
A numeric vector of locus lengths.
}
\description{
Calculates length of loci, possibly including gaps and data-defficient parts.
}
