% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainparams}
\alias{mainparams}
\title{Main parameters of STRUCTURE}
\format{
A character vector with the lines of 'mainparams' template provided with STRUCTURE.
}
\source{
https://web.stanford.edu/group/pritchardlab/structure.html
}
\usage{
mainparams
}
\description{
Template for main parameters file of STRUCTURE.
}
\keyword{datasets}
