% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_snps.R
\name{make_binary_snps}
\alias{make_binary_snps}
\title{Binary coding of biallelic SNPs.}
\usage{
make_binary_snps(
  snps,
  allele = "_[[:digit:]]$",
  counts = TRUE,
  format = c("snps", "structure"),
  center = FALSE,
  scale = FALSE
)
}
\arguments{
\item{snps}{single nucleotide polymorphisms in \code{matrix} or \code{DNAbin} format.}

\item{allele}{regular expression, allele identifier in the rownames of snps.}

\item{counts}{logical, whether to represent individual genotype in a single row by counts of '1' alleles.}

\item{format}{data format, either \code{"snps"} (default) or \code{"structure"}, diifers in treatment of ambiguous and missing data.}

\item{center}{logical, whether to represent genotypes as centered, i.e. with 0 zero for heterozygote and -1, 1 for homozygotes.}

\item{scale}{logical, whether to scale genotypes by the expected rate of genetic drift. If \code{center=FALSE} and \code{scale=TRUE},
the data are also centered so the column means are zero. By default, both \code{center} and \code{scale} are \code{FALSE}.}
}
\value{
A matrix with genotypes represented by \code{0} and \code{1} alleles or their counts (if \code{counts=TRUE}).
}
\description{
Converts biallelic SNPs to their binary representation.
}
