% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergentseq.R
\name{maxdist}
\alias{maxdist}
\title{Most distant cases.}
\usage{
maxdist(
  dst,
  n,
  phy = NULL,
  seq = NULL,
  outgroup = NULL,
  reserved = NULL,
  ord = NULL
)
}
\arguments{
\item{dst}{distance matrix, object of class \code{dist} or \code{matrix}, the latter is assumed to be square and symmetric.}

\item{n}{number of cases to retain.}

\item{phy}{an object of class \code{phylo} with phylogeny to be subset.}

\item{seq}{an object of class \code{DNAbin} or \code{matrix} with sequence alignment to be subset.}

\item{outgroup}{a character vector with outgroup names or their unique identifier(s),}

\item{reserved}{A character vector listing sequences that must be included in the selection.}

\item{model}{A character string specifying the nucleotide substitution model used to correct for multiple
mutations when calculating pairwise distences between sequences; deafult is \code{"raw"} (no correction).}
}
\value{
If \code{phy} or \code{seq} are supplied the function return subset phylogenetic tree or sequence alignment, respectively.
Otherwise, it returns a vector with IDs of cases to be retained.
}
\description{
Selection of a specified number of cases (e.g., sequences or tree tips), which are most distant from each other.
}
\details{
The algorithm operates on the supplied distance matrix. It can be used to subsample tree tips based on distances
between them (if \code{dst=cophenetic(phy)}) or sequences based on their mutual distance (if \code{dst=dist.dna(seq)}).
However, phylogenetic distances can be also used to subsample sequences and vice versa, or the distance matrice can be
of entirely different kind, e.g., matrix of geographical distances between individuals or genetic distances calculated
from another part of genome. Anyway, either \code{phy} or \code{seq} or none of them, but not both at the same time must be supplied.
}
