% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occmatrix.R
\name{occheatmap}
\alias{occheatmap}
\title{Occupancy heatmap.}
\usage{
occheatmap(occmat, device, file)
}
\arguments{
\item{occmat}{the output of \link{occmatrix}.}

\item{device}{either \code{"quartz"}, \code{"x11"} (= \code{"X11"}) or \code{"pdf"}.}

\item{file}{character strigm the name of pdf file if \code{device == "pdf"}.}
}
\description{
Creates a heatmap showing results of two-way occupancy analysis.
}
