% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occmatrix.R
\name{occmatrix}
\alias{occmatrix}
\title{Two-way occupancy.}
\usage{
occmatrix(loci, allele = "_[012]$")
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{matrix} or \code{DNAbin}.}

\item{allele}{regular expression, allele identifier in rownames of locus-specific alignments.}
}
\value{
A matrix with rows corresponding to individuals and columns to subsets of loci increasingly filtered
by their occupancy. It contains individual coverages in these subsets as calculated by \link{coverage}.
}
\description{
Calculates simultaneously occupancy of loci and coverage of individuals.
}
