% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opacity.R
\name{opacity}
\alias{opacity}
\title{Color opacity.}
\usage{
opacity(color, alpha)
}
\arguments{
\item{color}{a vector of colors (color names, hexadecimal strings or integers from \link[grDevices:palette]{grDevices::palette}).}

\item{alpha}{a numeric vector, degree of opacity (from 0 to 1), recycled if necessary.}

\item{returns}{A vector of modified colors.}

\item{details}{This function defines opacity of a color, 0 being fully transparent and 1 being fully visable.}
}
\description{
Makes colors opaque / transparent.
}
