% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_branches.R
\name{paint_branches}
\alias{paint_branches}
\title{Branch colors.}
\usage{
paint_branches(phy, info, color, bg = "black", nested = TRUE)
}
\arguments{
\item{phy}{an object of class \code{phylo}.}

\item{info}{a data frame (or matrix) with tip labels (1st column) and clade assignment (2nd column).
The classification need not to be comprehensive.}

\item{color}{named vector of colors, the names correspond to clade labels,
or a matrix or data frame with colors (1st column) and clade labels (2nd column).}

\item{bg}{color of unclassified ("background") branches.}

\item{nested}{logical, whether to paint branches belonging to the nested clades (if \code{TRUE})
or retain the color of the larger clade (if \code{FALSE}).}

\item{returns}{A vector of colors assigned to tree branches, ordered as in \code{phy$edge}.}
}
\description{
Assigns colors to branches depending which clade they belong to.
}
