% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{partition}
\alias{partition}
\title{Partition sequences.}
\usage{
partition(x, part = NULL, toupper = TRUE)
}
\arguments{
\item{x}{a sequence alignment in \code{matrix} or \code{DNAbin} format.}

\item{part}{either a two column matrix indicating starts and ends of loci in the concatenated aligments
(in .fasta or .nexus files) or a name of partition file formatted according to RAxML or MrBayes convention.}

\item{toupper}{logical, whether to export sequences in upper case letters.}

\item{missing}{a symbol representing missing nucleotides.}
}
\value{
A list of alignments in \code{matrix} or \code{DNAbin} format with \code{part} attribute defining locus boundaries.
}
\description{
Partitions multi-locus sequences.
}
