% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoord.R
\name{pcoord}
\alias{pcoord}
\title{Principal coordinate analysis.}
\usage{
pcoord(dst)
}
\arguments{
\item{d}{matrix of pairwise dissimilarities, possibly calculated from binary coded genotype data by \link{dissim.snps}.}
}
\value{
A list with components \code{values} and \code{vectors} with results of the spectral decomposition of
internally computed pairwise inter-individual covariance matrix.
}
\description{
Principal coordinate analysis, a.k.a. classical multidimensional scaling.
}
\details{
The function is an implementation of PCoA as described, e.g., in Numerical Ecology 3rd ed. (Legendre et al. 2012).
The spectral decomposition is performed by \link[base:eigen]{base::eigen}.
}
