% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomap.R
\name{phylomap}
\alias{phylomap}
\title{Phylogeographic map.}
\usage{
phylomap(
  info,
  color,
  subset,
  ind,
  lin,
  gps,
  border = "black",
  size = 4,
  lwd = 1,
  map = NULL,
  device,
  width = 7,
  height = 7,
  file = NULL,
  mai = 0.02,
  xlim,
  ylim,
  axes = TRUE,
  cex.axis = 1.25,
  legend = TRUE,
  expr
)
}
\arguments{
\item{info}{a data frame with individual identifier, classification into phylogeographic lineages
and geographical coordinates, whose names are partial matches of "latitude", "longitude" or "x", "y").}

\item{color}{a named vector of point colors, names must correspond to lineage labels.
If missing, a default palette is used.}

\item{subset}{logical, numeric or character; indicates subset of rows in info to be plotted. If logical, it must be
of the same length as it is the number of rows in info, if character, it contains individual IDs.}

\item{ind, lin}{optional character strings with column names including individual IDs (e.g. "ID")
and classification to phylogeographic lineage (e.g. "Species"), respectively. If not supplied,
these two columns are supposed to be the first and the second one, respectively.}

\item{gps}{optional data frame with geographical coordinates. It must contain the same individual identifier
as used in \code{info} and geographical coordinates named as explained for \code{info}. If \code{gps} is supplied, it supersedes
the information in \code{info}. If only a subset of IDs is matching those in \code{info}, only this subset is plotted.}

\item{border}{color of point borders. If \code{border == "auto"}, the borders are identical to point fillings.}

\item{size}{size of points specified as a percentage of the lesser of latitude / longitude ranges.}

\item{map}{map layer serving as background instead of country borders from \code{maps} package.}

\item{device}{either \code{"quartz"}, \code{"x11"} or \code{"pdf"} to indicate the type of graphical device.
If \code{NULL}, the objects are plotted into the current device.}

\item{width}{width of graphical device in inches.}

\item{height}{height of graphical device in inches.}

\item{file}{a name of pdf file if \code{device == "pdf"}.}

\item{mai}{size of outer margins in inches, recycled if necessary.}

\item{xlim, ylim}{Limits of x and y axes.}

\item{axes}{logical, whether to display axes.}

\item{cex.axis}{magnification of axis annotation.}

\item{legend}{logical, whether to show the legend (on a margin, with in-built parameters).}

\item{expr}{an expression (or a list of them) allowing to add other elements (e.g. customized legend).}
}
\description{
Creates a simple phylogeographic map.
}
