% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow.R
\name{plot.elbow}
\alias{plot.elbow}
\title{Elbow method.}
\usage{
\method{plot}{elbow}(x)
}
\arguments{
\item{x}{an object of class \code{elbow}, result of \link{elbow}.}
}
\description{
A plot method for the result of \link{elbow}.
}
