% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascertbias.R
\name{plot_ascertbias}
\alias{plot_ascertbias}
\title{Ascertainment bias plot.}
\usage{
plot_ascertbias(
  x,
  y,
  ascert,
  cumulative = TRUE,
  xlab = "",
  ylab = "",
  main = "",
  cex = 1.5,
  cex.lab = 1.5,
  cex.axis = 1.25,
  return = FALSE,
  device,
  file
)
}
\arguments{
\item{x}{string, a name of x-axis variable, e.g. 'occupancy'.}

\item{y}{string, a name of y-axis variable, e.g. 'mdist'.}

\item{ascert}{output of \link{ddrad_ascertbias}.}

\item{cumulative}{logical, whether to plot statistics for increasingly filtered sets of loci
rather than for different occupancy bins. The default is \code{TRUE}.}

\item{xlab, ylab, main}{labels of axes and the whole plot.}

\item{cex, cex.lab, cex.axis}{size of points, axis labels and tick labels.}

\item{return}{logical, whether to return a list with binned statistics (\code{classes}), their medians (\code{medians}),
break points between bins (\code{breaks}) and the numbers of loci contained in the bins (\code{nloci}).}

\item{device}{either \code{"quartz"}, \code{"x11"} (= \code{"X11"}) or \code{"pdf"}.}

\item{file}{character strigm the name of pdf file if \code{device == "pdf"}.}
}
\description{
Plots results of the ascertainment bias analysis.
}
