% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clades.R
\name{plot_clades}
\alias{plot_clades}
\title{Clades in the tree.}
\usage{
plot_clades(
  phy,
  clades,
  color,
  bg = "black",
  type = "phylogram",
  lines = TRUE,
  names = TRUE,
  order = FALSE,
  edge.color,
  edge.width = 3,
  scale = TRUE,
  tip_labels = TRUE,
  node_labels = FALSE,
  edge_labels = FALSE,
  root_label = NULL,
  anc_labels = TRUE,
  device,
  width = 7,
  height = 7,
  file = NULL,
  mai,
  extension = 8,
  line.pos = 0,
  line.over = 0.5,
  name.col = NULL,
  name.cex = 1.25,
  name.font = 1,
  name.las = 2,
  tip.lab.pos = 0.005,
  tip.lab.cex = 1,
  tip.lab.font = 3,
  node.lab.pos = 1,
  node.lab.cex = 1,
  node.lab.lwd = 1,
  node.lab.asp = 1,
  node.lab.col = "black",
  node.lab.bg = "white",
  node.text.pos = c(0.5, 0.5),
  node.text.cex = 1,
  node.text.font = 1,
  node.text.col = "black",
  edge.lab.pos = 0.5,
  edge.lab.cex = 1,
  edge.lab.lwd = 1,
  edge.lab.asp = 0.8,
  edge.lab.col = "black",
  edge.lab.bg = "white",
  edge.text.pos = c(0.5, 0.5),
  edge.text.cex = 1,
  edge.text.font = 1,
  edge.text.col = "black"
)
}
\arguments{
\item{phy}{an object of class \code{phylo}.}

\item{clades}{vector of clade labels (order must correspond to \code{phy$tip.label}) or a data frame (or matrix)
with tip labels (1st column) and clade assignment (2nd column). The classification need not to be comprehensive.}

\item{color}{named vector of colors. The names correspond to either clade labels or tip labels, the latter is
relevant especially if coloration is unlinked from clade assignment. It can be also a matrix or data frame
with labels (1st column) and colors (2nd column).}

\item{bg}{color of background (unclassified) branches.}

\item{type}{character, tree type, either \code{"phylogram"} (default) or \code{"cladogram"}.}

\item{lines}{character, tree type, either \code{"phylogram"} (default) or \code{"cladogram"}.}

\item{order}{either logical (whether to order the clades alphabetically) or character
(clade labels in the required order).}

\item{edge.color}{branch colors, vector of the same length as \code{nrow(phy$edge)}, if specified,
it supersedes \code{color} of branches, while clade lines and names are not affected.}

\item{edge.width}{numeric, branch width.}

\item{scale}{logical, whether to display scale bar as x-axis.}

\item{tip_labels}{vector of tip labels, their order corresponds to \code{phy$tip.label}.}

\item{node_labels}{a named vector of node labels to be displayed (names indicate node numbers) or logical,
with TRUE meaning to show \code{phy$node.label} (if available) and FALSE (default) suppresses display of node labels.}

\item{edge_labels}{a named vector of node labels to be displayed (names indicate edge order in \code{phy$edge}) or logical,
with TRUE meaning to use \code{phy$node.label} (if available) and FALSE (default) suppresses display of edge labels.}

\item{root_label}{a character string added as a root label to \code{node_labels}.}

\item{anc_labels}{logical, whether to show \code{node_labels} and \code{edge_labels} only for clade ancestors.}

\item{device}{either \code{"quartz"}, \code{"x11"} or \code{"pdf"} to indicate the type of graphical device.
If \code{NULL}, the objects are plotted into the current device.}

\item{width}{width of graphical device in inches.}

\item{height}{height of graphical device in inches.}

\item{file}{a name of pdf file if \code{device == "pdf"}.}

\item{mai}{size of outer margins in inches, recycled if necessary.}

\item{extension}{an amount of extension of axes expressed as percentage of the data range.}

\item{line.pos}{position of clade lines, negative values ~ shift to the left.}

\item{line.over}{overhang of clade lines as a proportion of single character height.}

\item{name.col}{color of clade names.}

\item{name.cex}{size of clade names.}

\item{name.font}{font of clade names.}

\item{name.las}{orientation of clades names relative to the axis (default perpendicular).}

\item{tip.lab.pos}{gap between tree tips and tip labels as a proportion of the maximum root-tip distance.}

\item{tip.lab.cex}{size of tip labels.}

\item{tip.lab.font}{font of tip labels.}

\item{node.lab.pos}{node label position as a proportion of the supporting branches.}

\item{node.lab.cex}{size of node label point.}

\item{node.lab.lwd}{width of node label point border.}

\item{node.lab.asp}{shape of node label, aspect ratio of an ellipse, default=1 means cricle.}

\item{node.lab.col}{color of node label point border.}

\item{node.lab.bg}{color of node label point background.}

\item{edge.lab.pos}{edge label position as a proportion of the branch from its start (= root side).}

\item{edge.lab.cex}{size of edge label point.}

\item{edge.lab.lwd}{width of edge label point border.}

\item{edge.lab.asp}{shape of edge label, aspect ratio of an ellipse, default=1 means cricle.}

\item{edge.lab.col}{color of edge label point border.}

\item{edge.lab.bg}{color of edge label point background.}

\item{text.lab.pos}{adjustment of text position in node / edge label (adj of 'text').}

\item{text.lab.cex}{size of text in node / edge label.}

\item{text.lab.font}{font of text in node / edge label.}

\item{text.lab.col}{color of text in node / edge label.}
}
\description{
Plots phylogenetic tree while highlighting defined clades.
}
