% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phylomap.R
\name{plot_phylomap}
\alias{plot_phylomap}
\title{Phylogeographic map.}
\usage{
plot_phylomap(
  coords,
  class,
  ind,
  group,
  color,
  border = "black",
  size = 4,
  lwd = 1,
  map = NULL,
  device,
  width = 7,
  height = 7,
  file = NULL,
  mai = 0.02,
  xlim,
  ylim,
  axes = TRUE,
  cex.axis = 1.25,
  expr
)
}
\arguments{
\item{coords}{a data frame with individual identifier (e.g., \code{ID} column) and geographical coordinates,
whose names are partial matches of "latitude", "longitude" or "x", "y").}

\item{class}{a data frame with individual identifier (1st column) and classification into groups (2nd column)
supposed to be phylogenetic lineages, species, clades...}

\item{ind}{a name of individual identifier (e.g. "ID").}

\item{group}{a name of group identifier (e.g. "Species").}

\item{color}{a named vector of point colors, names must correspond to group labels.
If missing a default palette is used.}

\item{border}{color of point borders. If \code{border == "auto"}, the borders are identical to point fillings.}

\item{size}{size of points specified as a percentage of the lesser of latitude / longitude ranges.}

\item{map}{map layer serving as background instead of country borders from \code{maps} package.}

\item{device}{either \code{"quartz"}, \code{"x11"} or \code{"pdf"} to indicate the type of graphical device.
If \code{NULL}, the objects are plotted into the current device.}

\item{width}{width of graphical device in inches.}

\item{height}{height of graphical device in inches.}

\item{file}{a name of pdf file if \code{device == "pdf"}.}

\item{mai}{size of outer margins in inches, recycled if necessary.}

\item{xlim, ylim}{Limits of x and y axes.}

\item{axes}{logical, whether to display axes.}

\item{cex.axis}{magnification of axis annotation.}

\item{expr}{an expression (or a list of them) allowing to add other elements (e.g. legend).}
}
\description{
Creates a simple phylogeographic map.
}
