% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_str_output.R
\name{plot_str_output}
\alias{plot_str_output}
\title{Plotting STRUCTURE output file.}
\usage{
plot_str_output(
  out,
  palette,
  interval = FALSE,
  ord = TRUE,
  device,
  file,
  width = 10,
  height = 5,
  mai = NULL,
  show.ind.label = FALSE,
  side.ind.label = 1,
  ylab = "Ancestry proportions",
  main = NULL,
  line.axis = 0.25,
  cex.axis = 1,
  cex.lab = 1,
  cex.main = 1,
  font.lab = 1,
  expr,
  ...
)
}
\arguments{
\item{out}{data frame produced by \link{read_str_output}.}

\item{palette}{character, colors assigned to the clusters in their consecutive order.}

\item{interval}{logical, whether to display width of probability interval by color transparency.}

\item{ord}{character or logical, order of individuals in the plot (from left to right).
If \code{TRUE} (default) it is determined internally, if \code{FALSE} it is ignored.}

\item{device}{either \code{"quartz"}, \code{"x11"} (= \code{"X11"}) or \code{"pdf"}.
If \code{NULL}, the objects are plotted into the current device.}

\item{file}{character strigm the name of pdf file if \code{device == "pdf"}.}

\item{width}{width of the device in inches.}

\item{height}{height of the device in inches.}

\item{mai}{numeric vector, sizes of outer margins (bottom, left, top, right) in inches.}

\item{show.ind.label}{logical, whether to display individual labels.}

\item{side.ind.label}{numeric, where to display individual labels (default is \code{1}, i.e., below the plot).}

\item{cex.axis}{size of tick labels.}

\item{cex.lab}{size of tick labels.}

\item{expr}{expression allowing to add other elements (e.g. legend).}
}
\description{
Creates barplot of STRUCTURE output.
}
