% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mrbayes_input.R
\name{prepare_mbblock}
\alias{prepare_mbblock}
\title{MrBayes nexus block.}
\usage{
prepare_mbblock(part, mcmc, sumt, priors = NULL, topology = NULL)
}
\arguments{
\item{part}{either a two column matrix indicating starts and ends of loci in the concatenated aligments
(in .fasta or .nexus files) or a name of partition file formatted according to RAxML or MrBayes convention.}

\item{mcmc}{a vector of MCMC parameters, named by the parameter names used by MrBayes,
the parameters not included are given default values (see Details).}

\item{sumt}{a vector of sumt command parameters, named by the parameter names used by MrBayes,
the parameters not included are given default values (see Details). If \code{FALSE}, it is omitted.}

\item{priors}{not implemented yet}

\item{topology}{topology to be fixed in MrBayes analysis, specified in any way acceptable to \link{read_tree}.}
}
\description{
Assembles the nexus block of MrBayes commands.
}
\details{
The \code{mcmc} defaults are: ngen=2500000, samplefreq=1000, relburnin="yes", burninfrac=0.1, savebrlens="yes", nchains=4, nruns=4.
The \code{sumt} defaults are: contype="allcompat", minpartfreq="0.0", calctreeprobs="no", showtreeprobs="no", hpd="yes", relburnin="yes",
burninfrac=0.1.
}
