% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fasta.R
\name{read.fasta}
\alias{read.fasta}
\title{Reading fasta file.}
\usage{
read.fasta(file)
}
\arguments{
\item{file}{character string, the name of the input fasta file.}
}
\value{
A matrix of mode \code{character} with nucleotides represented by upper case letters.
The coding of gap, missing or ambiguous data is left as it was in the input file.
}
\description{
A wrapper for \link[ape:read.dna]{ape::read.dna} with options \code{format="fasta"}, \code{as.matrix=TRUE} and \code{as.character=TRUE}.
}
