% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_snps.R
\name{read_binary}
\alias{read_binary}
\title{Reading file with SNP genotypes.}
\usage{
read_binary(file)
}
\arguments{
\item{file}{character string, the file name.}
}
\value{
A matrix with binary coded SNP genotypes.
}
\description{
Read a tab-delimited file with binary coded SNP genotypes.
}
