% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.R
\name{read_clumpak_output}
\alias{read_clumpak_output}
\title{Reading CLUMPAK output file.}
\usage{
read_clumpak_output(file, ind = NULL)
}
\arguments{
\item{file}{character string, name of the output file.}
}
\value{
A matrix with estimated ancestry proportions.
}
\description{
Reads CLUMPAK output data file.
}
\details{
So far it works only for admixture model with no POPINFO prior.
}
