% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seq.R
\name{read_fasta}
\alias{read_fasta}
\title{Importing fasta files.}
\usage{
read_fasta(
  file,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part = NULL,
  rm.empty = FALSE,
  nlines,
  indiv,
  interleaved
)
}
\arguments{
\item{file}{character string, the name(s) of file(s) to be imported (each file corresponds to a single locus).}

\item{nloci}{the number of loci to be imported (the default \code{-1L} means all of them).}

\item{class}{character, the output class, either \code{"matrix"} or \code{"DNAbin"}.}

\item{toupper}{logical, whether to represent nucleotides by upper case letters.}

\item{part}{it can be a two column matrix indicating starts and ends of loci in the concatenated aligments
or a name of partition file formatted according to RAxML or MrBayes convention.}

\item{rm.empty}{logical, whether to remove empty (e.g. all "-" or all "N") sequences.}

\item{nlines}{not relevant, just for compatibility with \link{read_seq}.}

\item{indiv}{not relevant, just for compatibility with \link{read_seq}.}

\item{interleaved}{not relevant, just for compatibility with \link{read_seq}.}
}
\value{
A list of locus-specific alignments.
}
\description{
Importing sequence data from files in .fasta format.
}
