% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seq.R
\name{read_ipyrad}
\alias{read_ipyrad}
\title{Reading ipyrad output.}
\usage{
read_ipyrad(
  file,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part,
  rm.empty,
  nlines = -1L,
  indiv,
  interleaved
)
}
\arguments{
\item{file}{character string, the name of the .alleles file.}

\item{nloci}{the number of loci to be read in (the default \code{-1L} means all of them).}

\item{class}{character, the output class, either \code{"matrix"} or \code{"DNAbin"}.}

\item{toupper}{logical, whether to represent sequences in upper case letters.}

\item{part}{not relevant, just for compatibility with \link{read_seq}.}

\item{rm.empty}{not relevant, just for compatibility with \link{read_seq}.}

\item{nlines}{the number of lines to be read in (the default \code{-1L} means all of them).
An incomplete locus at the end of the lines is removed.}

\item{indiv}{not relevant, just for compatibility with \link{read_seq}.}

\item{interleaved}{not relevant, just for compatibility with \link{read_seq}.}
}
\value{
A list of locus-specific sequence alignments. If the assembly of RAD loci was by mapping
to a reference genome, the output has an attribute \code{bedtable}.
}
\description{
Reading sequence data from ipyrad .alleles files.
}
