% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seq.R
\name{read_nexus}
\alias{read_nexus}
\title{Reading nexus files.}
\usage{
read_nexus(
  file,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part = NULL,
  rm.empty = FALSE,
  nlines,
  indiv,
  interleaved
)
}
\arguments{
\item{file}{character string, the name of .nexus file.}

\item{nloci}{the number of loci to be read in (the default \code{-1L} means all of them).}

\item{class}{character, the output class, either 'matrix' or 'DNAbin'.}

\item{toupper}{logical, whether to represent nucleotides by upper case letters.}

\item{part}{it can be a two column matrix indicating starts and ends of loci in the concatenated aligments
or a name of partition file formatted according to RAxML or MrBayes convention. It can be overridden
by the nexus block defining character sets, if present within the sequence file.}

\item{rm.empty}{logical, whether to remove empty (e.g. all "-" or all "N") sequences.}

\item{nlines}{not relevant, just for compatibility with \link{read_seq}.}

\item{indiv}{not relevant, just for compatibility with \link{read_seq}.}

\item{interleaved}{not relevant, currently just for compatibility with \link{read_seq}.}
}
\value{
A list of locus-specific alignments. If genomic coordinates are included in a "bedtable" block of the file
they are exported in .bed format as an attribute \code{bedtable}.
}
\description{
Reading sequence data from files in .nexus format.
}
