% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_part.R
\name{read_part}
\alias{read_part}
\title{Read partition file.}
\usage{
read_part(file, param = FALSE, name = NULL)
}
\arguments{
\item{file}{a name of the partition file.}

\item{param}{logical, whether to read also estimates of substitution model parameters.}

\item{name}{character string with locus name, relevant when reading in standard output of IQ-TREE
(rather than a partition file it creates).}
}
\value{
A data frame with columns for the partition name, start, end, codon position
and possibly for components of nucleotide substitution models and estimates of their parameters.
Every row in the data frame corresponds to a single elementary partition, which can be merged with another one,
as indicated by the partition name.
}
\description{
Reads nexus or RAxML-formatted partition files, including files created by ModelFinder tool of IQ-TREE software.
}
