% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seq.R
\name{read_seq}
\alias{read_seq}
\title{Importing sequence data.}
\usage{
read_seq(
  file,
  format,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part = NULL,
  rm.empty = FALSE,
  nlines = -1L,
  indiv = NULL,
  interleaved = TRUE
)
}
\arguments{
\item{file}{character string, the name(s) of file(s) to be imported.}

\item{format}{character, the format of the file(s); it can be \code{"ipyrad"}, \code{"stacks"}, \code{"nexus"}, \code{"fasta"} or \code{"phylip"}.}

\item{nloci}{the number of loci to be imported (the default \code{-1L} means all of them)}

\item{class}{character, the output class, either \code{"matrix"} or \code{"DNAbin"}.}

\item{toupper}{logical, whether to represent nucleotides by upper case letters.}

\item{part}{either a two column matrix indicating starts and ends of loci in the concatenated aligments
(in .fasta or .nexus files) or a name of partition file formatted according to RAxML or MrBayes convention.
In \link{read_nexus} it is overridden by the nexus block defining character sets, if present within the sequence file.}

\item{rm.empty}{logical, whether to remove empty (e.g. all "-" or all "N") sequences.}

\item{nlines}{the number of lines to be imported (the default \code{-1L} means all of them;
relevent for formats \code{"ipyrad"} and \code{"stacks"}. An incomplete locus at the end of the imported lines is removed.}

\item{indiv}{information about individual names, relevant for the format \code{"stacks"}. It may be \code{"pop"}
(individuals correspond to populations in 'Stacks'), a vector of individual names in an order corresponding
to sample numbers in Stacks or \code{NULL} (default), which makes individuals being numbered as in Stacks.}

\item{interleaved}{Logical, whether in interleaved (not sequential) format; currently only for \link{read_phylip}.}
}
\value{
A list of locus-specific alignments. If genomic coordinates are included (in ipyrad or Stacks output
or in a "bedtable" block of .nexus file), they are exported in .bed format as an attribute \code{bedtable}.
}
\description{
The functions importing sequence data from various file formats.

\code{read_seq} is a wrapper for the following functions.

\code{read_nexus} from .nexus file, loci defined by an appropriate nexus block or an optional argument.

\code{read_fasta} from .fasta file, loci contained in different files or defined by an optional argument.

\code{read_phylip} from .phy file, loci contained in different files or defined by an optional argument.

\code{read_ipyrad} from .alleles file of ipyrad RADseq assembler.

\code{read_stacks} from .alleles.fas file of stacks RADseq assembler.
}
