% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seq.R
\name{read_stacks}
\alias{read_stacks}
\title{Importing stacks output.}
\usage{
read_stacks(
  file,
  nloci = -1L,
  class = "matrix",
  toupper = TRUE,
  part,
  rm.empty,
  nlines = -1L,
  indiv = NULL,
  interleaved
)
}
\arguments{
\item{file}{character string, the name of the .alleles.fas file.}

\item{nloci}{the number of loci to be imported (the default \code{-1L} means all of them).}

\item{class}{character, the output class, either \code{"matrix"} or \code{"DNAbin"}.}

\item{toupper}{logical, whether to import sequences in upper case letters.}

\item{part}{not relevant, just for compatibility with \link{read_seq}.}

\item{rm.empty}{not relevant, just for compatibility with \link{read_seq}.}

\item{nlines}{the number of lines to be imported (the default \code{-1L} means all of them).
An incomplete locus at the end of the imported lines is removed.}

\item{indiv}{information about individual names. It may be \code{"pop"} (individuals correspond to populations in 'Stacks'),
a vector of individual names in an order corresponding to sample numbers in Stacks or \code{NULL} (default),
which makes individuals being numbered as in Stacks.}

\item{interleaved}{not relevant, just for compatibility with \link{read_seq}.}
}
\value{
A list of locus-specific sequence alignments. If the assembly of RAD loci was by mapping
to a reference genome, the output has an attribute \code{bedtable}.
}
\description{
Importing sequence data from stacks . alleles.fas files.
}
