% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.R
\name{read_str_data}
\alias{read_str_data}
\title{Reading STRUCTURE data file.}
\usage{
read_str_data(
  file,
  label = TRUE,
  popdata = FALSE,
  popflag = FALSE,
  locdata = FALSE,
  markernames = TRUE,
  recessive = FALSE,
  linkage = FALSE,
  phase = FALSE,
  ploidy = 2,
  allele
)
}
\arguments{
\item{file}{character string, name of the data file.}

\item{label}{logical, whether the first column contains individual labels, default is \code{TRUE}.}

\item{popdata, popflag, locdata}{logical, whether the initial columns contain information on population of origin (\code{popdata}),
whether to use it in clustering (\code{popflag}) and information on "locality", i.e. the clustering affecting prior (\code{locdata}).}

\item{markernames}{logical, whether the first row contains marker (=locus) names, default is \code{TRUE}.}

\item{recessive, linkage}{logical, whether the initial rows contain information which allele is recessive (\code{recessive}),
and what are genomic distances separating the markers (\code{linkage }).}

\item{phase}{logical, whether rows with individual genotypes are alternated by rows with phase information.}

\item{ploidy}{numeric, the ploidy, i.e., the number of rows per individual genotype, default is 2.}

\item{allele}{optional character vector with identifiers used to distinguish individual alleles, e.g., \code{c("_0", "_1")},
used only if there is any phase information.}
}
\value{
A matrix with genotypes, associated with an attributes \code{pops} (population information from the first columns),
\code{markers} (information about markers the first rows) and \code{phase} with information about phasing of individual genotypes.
}
\description{
Reads STRUCTURE input data file.
}
\details{
The way STRUCTURE data files are formatted requires explicit specification, which rows and columns are included.
By default, the function assumes the file to contain individual and marker names (\code{label=TRUE} and \code{markernames=TRUE}),
but no other information. The function can handle data of arbitrary ploidy, if it does not vary across individuals.
Currently, it does not support formatting with one row per individual diploid genotype, although this can be by-passed
by setting \code{ploidy=1}.
}
