% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.R
\name{read_str_output}
\alias{read_str_output}
\title{Reading STRUCTURE output file.}
\usage{
read_str_output(file)
}
\arguments{
\item{file}{character string, name of the output file.}
}
\value{
A matrix with estimated ancestry proportions and their credibility intervals.
}
\description{
Reads STRUCTURE output data file.
}
\details{
So far it works only for admixture model with no POPINFO prior.
}
