% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree.R
\name{read_tree}
\alias{read_tree}
\title{Read trees.}
\usage{
read_tree(tree, thin = 1)
}
\arguments{
\item{tree}{a character string with a tree written in newick format or a file name,
or an object of class \code{phylo} or \code{multiPhylo}.}

\item{thin}{a proportion of mcmc steps to be retained or a subsampling frequency (if \code{thin > 1}).}
}
\value{
An object of class \code{phylo} or \code{multiPhylo}.
}
\description{
Reads phylogenetic trees from a newick text string or a file written in newick or nexus format.
}
