% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_repeats.R
\name{remove_repeats}
\alias{remove_repeats}
\title{Removal of repeats.}
\usage{
remove_repeats(loci, minn = 5, kmer = 5, delete = FALSE)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{minn}{the minimum number of repeats considered as making the sequence repetitive.}

\item{kmer}{the maximum size of k-mers considered as elementary repeat motifs.}

\item{delete}{logical, whether to delete gap-only columns that appear due to removal of repetitions.}
}
\value{
A modified list of locus-specific sequence alignments in the same format as was the input.
}
\description{
Mark repetitive parts of sequences as missing data.
}
