% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_tree.R
\name{reorder_clades}
\alias{reorder_clades}
\title{Reorder clades.}
\usage{
reorder_clades(phy, clades, cladeorder)
}
\arguments{
\item{phy}{an object of class \code{phylo} with the input tree.}

\item{clades}{a data frame listing tip labels (1st column) and clade labels (2nd column).
The specified clades must be monophyletic in the tree, but may cover only a part of it.}

\item{cladeorder}{a vector giving clade labels in the desired order.}
}
\value{
Rotated tree as a \code{phylo} object.
}
\description{
Reorders clades in the tree, so they are plotted in the specified order.
}
