% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_tree.R
\name{reorder_tree}
\alias{reorder_tree}
\title{Reorder tree tips.}
\usage{
reorder_tree(phy, tiporder)
}
\arguments{
\item{phy}{an object of class \code{phylo} with the input tree.}

\item{tiporder}{a vector giving tip labels in the desired order or a \code{phylo} object from which
the tip order is taken. Partial constraint on the tip order is accepted.}
}
\value{
Rotated tree as a \code{phylo} object.
}
\description{
Rotates clades to achieve specified ordering of tree tips (if possible, given the tree topology).
}
\details{
The function uses iteratively \link[ape:rotate]{ape::rotate} and then exports the rotated \code{phy} object using
\code{ape::read.tree(text=ape::write.tree(phy))}. If the input contains components \code{edge.label} or \code{edge.color},
they are retained and their elements reordered as necessary.
}
