% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_raxml.R
\name{run_epa}
\alias{run_epa}
\title{Evolutionary placement analysis.}
\usage{
run_epa(
  command,
  args,
  prefix,
  outdir,
  retain = "\\\\.jplace$",
  rename = TRUE,
  stdout = "",
  stderr = ""
)
}
\arguments{
\item{command}{a character string with the system command. If not supplied, either "raxmlHPC-SSE3" or "raxmlHPC.exe" is used,
depending on the platform. Also the path to directory with the appropriate binary file can be specified
instead of the command itself.}

\item{args}{as in \link{run_raxml}, but not only "-s" but also "-t" specification is obligatory and "-f" is fixed to "v".
For defaults of other parameters, check \code{Details}.}

\item{prefix}{a character string to be included in names of all output files. It has priority over the specification
in \code{args} argument. If not specified, whether through \code{prefix} or \code{args}, the input file name without extension is used.}

\item{outdir}{a character string with the path to output directory. It has priority over the path included in \code{prefix}.
If not specified in any way, it is assumed to be the same as the working directory.}

\item{retain}{regular expressions with identifiers of outputs to be retained. The .jplace output is always retained, but
others can be retained as well (specify \code{retain="^RAxML_"} to retain all fo them).
changed, the other outputs than the ".jplace" file are deleted.}

\item{rename}{if \code{TRUE} (default) the name of .jplace output file is changed to \code{paste0(prefix, ".jplace")}.
Alternatively, the name can be specified by a character string or left as it is (\code{rename=FALSE}).}

\item{stdout, stderr}{where output to ‘stdout’ or ‘stderr’ should be sent, as in \link[base:system2]{base::system2}.}
}
\description{
Runs evolutionary placement analysis (EPA) under RAxML, equivalent to \code{run_raxml} with appropriately set arguments.
}
\details{
The minimum \code{args} contains "-s" and the name of the input file as well as "-t" and the name of the tree file.
The input directory is either retrieved from the input name (if it contains a path) or assummed to be the working directory.
Internally, the function always adds \code{c("-f", "v")} to \code{args} and if not specified otherwise by the user, also adds defaults
for EPA parameters (\code{"--epa-keep-placements=100"} and \code{"--epa-prob-threshold=0.01"}) and for convenience also "--HKY85"
as default substitution model. For the reference to RAxML commands, consult RAxML manual at
https://cme.h-its.org/exelixis/resource/download/NewManual.pdf.
}
