% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_raxml.R
\name{run_raxml}
\alias{run_raxml}
\title{RAxML analysis.}
\usage{
run_raxml(
  command,
  args,
  prefix,
  outdir,
  retain,
  stdout = "",
  stderr = "",
  seed = 123
)
}
\arguments{
\item{command}{a character string with the system command. If not supplied, either "raxmlHPC-SSE3" or "raxmlHPC.exe" is used,
depending on the platform. Also the path to directory with the appropriate binary file can be specified
instead of the command itself.}

\item{args}{a character vector in the form accepted by \code{args} argument of \link[base:system2]{base::system2} or a data frame with arguments
to \code{command} (argument names in the 1st column, values in the 2nd column) or a name of a tab-delimited file with the arguments.}

\item{prefix}{a character string to be included in names of all output files. It has priority over the specification
in \code{args} argument. If not specified, whether through \code{prefix} or \code{args}, the input file name without extension is used.}

\item{outdir}{a character string with the path to output directory. It has priority over the path included in \code{prefix}.
If not specified in any way, it is assumed to be the same as the working directory.}

\item{retain}{regular expressions with identifiers of outputs to be retained. If specified, the other outputs are deleted.}

\item{stdout, stderr}{where output to ‘stdout’ or ‘stderr’ should be sent, as in \link[base:system2]{base::system2}.}

\item{seed}{random seed for the analysis ("-p" argument of RAxML). It has priority over the specification in \code{args}.}
}
\description{
Runs RAxML analysis specified by the arguments, including pre- and post-processing of the files.
}
\details{
The minimum \code{args} contains "-s" and the name of the input file. The subsotution model ("-m") is "GTRGAMMA" by default.
The input directory is either retrieved from the input name (if it contains a path) or assummed to be the working directory.
The argument names must be supplied  with the leading dash ("-"). If \code{args} is a data frame, the arguments with double dash ("--")
have \code{NA} in the 2nd column.
For the reference to RAxML commands, consult https://cme.h-its.org/exelixis/web/software/raxml/hands_on.html
or RAxML manual at https://cme.h-its.org/exelixis/resource/download/NewManual.pdf.
}
