% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_snps.R
\name{scale_binary_snps}
\alias{scale_binary_snps}
\title{Scale SNP genotypes.}
\usage{
scale_binary_snps(b, center = TRUE, counts = TRUE)
}
\arguments{
\item{b}{matrix, the output of \link{make_binary_snps}.}

\item{center}{logical, whether to apply centering (default is \code{TRUE}).}

\item{counts}{logical, whether individual genotype is represented in a single row by counts of '1' alleles.}
}
\value{
A matrix with centered binary coded genotypes.
}
\description{
Scales and possibly also centers binary coded SNP genotypes. The centering is different from \link{center_binary_snps},
as it just makes all column means to be zero.
}
