% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqlen.R
\name{seqlen}
\alias{seqlen}
\title{Sequence lengths.}
\usage{
seqlen(input)
}
\arguments{
\item{input}{a character string, the name of file in .fasta or .phylip format
or a sequence alignment in \code{DNAbin} or \code{matrix} format.}
}
\value{
An numeric vector with sequence lengths.
}
\description{
Calculates lengths of sequences in the input alignment.
}
