% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyseq.R
\name{seqmatches}
\alias{seqmatches}
\title{Sequence matches.}
\usage{
seqmatches(taxa, haps, na.rm = FALSE, retain = TRUE, col.names)
}
\arguments{
\item{taxa}{a data frame with information about classification of haplotypes to taxa (species, clades, lineages)
or a name of file with such classification. Haplotype names are in the 1st column, taxa in the 2nd column.}

\item{haps}{either a data frame with assignment of sequences to haplotypes (as produced by \link{haplotypes})
or a name of file with such an assignment. Sequence names are in the 1st column, haplotype names in the 2nd column.}

\item{na.rm}{logical, whether to omit sequences that could not be classified into taxa.}

\item{retain}{logical, whether to retain all columns of \code{taxa} data frame or just the first two of them.}

\item{col.names}{optional character vector, column names of the output data frame.}
}
\value{
A data frame with similar to \code{info} input, but having in the 1st column sequence names
and in the 2nd column assignment to taxa, based on the correspondence of haplotype names.
}
\description{
Classify sequences into taxa based on their identity to already classified ones.
}
