% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps_utils.R
\name{single_snps}
\alias{single_snps}
\title{Single SNP per locus.}
\usage{
single_snps(
  snps,
  pis = FALSE,
  nalleles = 4,
  minmiss = FALSE,
  drop = TRUE,
  seed = sample(1e+08, 1)
)
}
\arguments{
\item{snps}{A list of locus-specific SNP matrices, as produced by \link{get_snps}.}

\item{pis}{Logical, whether to retain only parsimony-informative sites.}

\item{nalleles}{Numeric, the maximum no. of alleles per variable site.}

\item{minmiss}{Logical, whether to select preferentially SNPs with the minimum of mossing data.}

\item{drop}{Logical, whether to drop out SNPs not complying the criteria (default is \code{TRUE}),
or whether to give them lower priority (if \code{FALSE}).}

\item{seed}{A seed for random sampling, its specification enforces repeatability of the random selection.}
}
\value{
A list of locus-specific SNP matrices, but with just one site (column) per locus. Invariant loci a discarded.
}
\description{
Selecting one SNP per locus, randomly within constraints given by the function arguments.
}
