% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_snps.R
\name{subset_binary_snps}
\alias{subset_binary_snps}
\title{Subset of SNPs.}
\usage{
subset_binary_snps(b, rows = NULL, cols = NULL)
}
\arguments{
\item{b}{matrix, the output of \link{make_binary_snps}.}

\item{rows, }{cols indices of rows and columns to be retained.}
}
\value{
A matrix with subset of binary coded genotypes.
}
\description{
Subsets binary coded SNPs.
}
