% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_utils.R
\name{subset_indiv}
\alias{subset_indiv}
\title{Subsets of individuals.}
\usage{
subset_indiv(loci, subset, allele = NULL, invert = FALSE)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{subset}{individual identifiers.}

\item{allele}{a regular expression defining an allele identifier (allows to distinguish individual IDs).}

\item{invert}{logical, whether to remove rather than retain the specified sequences.}
}
\value{
A numeric vector of locus-specific occupancies.
}
\description{
Makes a subset of individuals acroiss loci.
}
