% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_utils.R
\name{subset_loci}
\alias{subset_loci}
\title{Subsets of loci.}
\usage{
subset_loci(loci, subset = NULL, n, seed = sample(1e+08, 1))
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{subset}{a vector with names or numeric indices of loci or logical values.}

\item{n}{a number of loci to be randomly selected.}

\item{seed}{numeric, a seed for random sampling.}
}
\value{
A numeric vector of locus-specific occupancies.
}
\description{
Makes a subset of loci.
}
