% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seq.R
\name{substrings}
\alias{substrings}
\title{Substring extraction.}
\usage{
substrings(i, j, x, n)
}
\arguments{
\item{i}{numeric, start.}

\item{j}{numeric, end.}

\item{x}{the full character string.}

\item{n}{character, locus name.}
}
\value{
The specified character substring.
}
\description{
Extracts a part of charater string.
}
\details{
Internal function of \link{read_seq}.
}
