% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytree.R
\name{summarytree}
\alias{summarytree}
\title{Summary tree.}
\usage{
summarytree(
  trees,
  type = c("MAP", "MCC", "CON"),
  file,
  brlen = TRUE,
  method = "mean",
  strict = FALSE,
  thin = 1,
  digits = 2
)
}
\arguments{
\item{trees}{a posterior sample of rooted trees in a \code{multiPhylo} object or a list of \code{phylo} objects
or a name of file with multiple trees in a format accepted by \link{read_tree}.}

\item{type}{the type of summary tree, either \code{"MAP"} (maximum aposteriori, default),
\code{"MCC"} (maximum clade credibility) or \code{"CON"} (majority consensus), see Details for explanantion.}

\item{file}{a name of the output file.}

\item{brlen}{logical, whether to estimate also branch lengths from the sample.}

\item{method}{string, name of the function to calculate branch lengths.}

\item{strict}{logical, whether to estimate branch lengths only from the trees, whose topology is strictly
identical with that of a summary tree.}

\item{thin}{a proportion of mcmc steps to be retained or a subsampling frequency (if \code{thin > 1}).}

\item{digits}{the number of digits in PP values.}
}
\value{
An object of class \code{phylo} with tree whose node labels indicate posterior probabilities of the clades.
}
\description{
Finds a summary tree for a posterior sample from a Bayesian phylogenetic inference.
}
\details{
The maximum a posteriori tree is the most frequently sampled tree.
The maximum clade credibility tree scores every sampled tree by the product of PPs of its constituent clades.
The majority consensus tree adds the most supported clades unless they overlap (they are in conflict)
with any of the clades that were already included.
}
