% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_traces.R
\name{thinning}
\alias{thinning}
\title{Thinning posterior samples.}
\usage{
thinning(x, freq = 1)
}
\arguments{
\item{x}{an object of class \code{multiPhylo} or a data frame.}

\item{freq}{subsampling frequency or a proportion of samples to be retained.}
}
\value{
An object of class \code{multiPhylo} or a data frame containing reduced posterior sample.
}
\description{
Subsamples posterior samples so every i-th value is retaind.
}
