% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{topology}
\alias{topology}
\title{Tree topology.}
\usage{
topology(tree, tiporder = NULL)
}
\arguments{
\item{tiporder}{a vector giving tip labels in the desired order or a \code{phylo} object from which
the tip order is taken. Partial constraint on the tip order is accepted.}

\item{phy}{an object of class \code{phylo} or \code{multiPhylo} or anything accepted by \link{read_tree}.}
}
\value{
A character string with tree topology (no branch lengths) in Newick format.
}
\description{
Writes tree topology in Newick format.
}
