% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trubetskoy}
\alias{trubetskoy}
\title{Color palette 'trubetskoy'}
\format{
A named vector of mode character specifying colors in HEX format.
}
\source{
https://sashat.me/2017/01/11/list-of-20-simple-distinct-colors
}
\usage{
trubetskoy
}
\description{
Twenty distinct colors designed by B. Trubetskoy.
}
\keyword{datasets}
