% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fasta.R
\name{write.fasta}
\alias{write.fasta}
\title{Writing fasta file.}
\usage{
write.fasta(seq, file)
}
\arguments{
\item{seq}{a matrix of mode \code{character} or an object of class \code{DNAbin}.}

\item{file}{character string, the ame of the output fasta file.}
}
\description{
A wrapper for \link[ape:write.dna]{ape::write.dna} with options \code{format="fasta"}, \code{as.matrix=TRUE} and \code{as.character=TRUE}.
}
\details{
It writes nucleotide in upper case, irrespective of their original coding.
The coding of gap, missing or ambiguous data is left as it was in the input object.
}
