% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_snps.R
\name{write_binary}
\alias{write_binary}
\title{Writing file with SNP genotypes.}
\usage{
write_binary(b, file)
}
\arguments{
\item{b}{matrix, the output of \link{make_binary_snps}.}

\item{file}{character string, the file name.}
}
\description{
Writes a tab-delimited file with binary coded SNP genotypes.
}
