% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_part.R
\name{write_codon_part}
\alias{write_codon_part}
\title{Codon partition scheme.}
\usage{
write_codon_part(
  scheme,
  file,
  last,
  first = 1,
  name = "pos",
  format = c("raxml", "nexus")
)
}
\arguments{
\item{scheme}{a character string symbolically representing partitioning scheme.}

\item{file}{a name of the partition file, if missing, the output is printed to console.}

\item{last}{numeric, number of sites in the alignement.}

\item{first}{numeric, position of the first nucleotide (within a longer genomic sequence).}

\item{name}{a character string, label(s) used to name partitions, if necessary, it is recycled and given the codon position numbers.}

\item{format}{a character string, either \code{"raxml"} or \code{"nexus"}.}
}
\description{
Writes nexus or RAxML-formatted partition file for a single codon-partitioned locus.
}
