% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mrbayes_input.R
\name{write_mrbayes_input}
\alias{write_mrbayes_input}
\title{MrBayes input file.}
\usage{
write_mrbayes_input(
  seq,
  file,
  part,
  mcmc = NULL,
  sumt = NULL,
  prefix = NULL,
  priors = NULL,
  topology = NULL,
  interleaved = FALSE,
  beagle = FALSE
)
}
\arguments{
\item{seq}{a sequence alignment in \code{matrix} or \code{DNAbin} format or a name of the file with the alignment.}

\item{file}{a name of the MrBayes file input to be created by the function.}

\item{part}{either a two column matrix indicating starts and ends of loci in the concatenated aligments
(in .fasta or .nexus files) or a name of partition file formatted according to RAxML or MrBayes convention.}

\item{mcmc}{a vector of MCMC parameters, named by the parameter names used by MrBayes,
the parameters not included are given default values (see Details).}

\item{sumt}{a vector of sumt command parameters, named by the parameter names used by MrBayes,
the parameters not included are given default values (see Details). If \code{FALSE}, it is omitted.}

\item{prefix}{a stem of the name}

\item{priors}{not implemented yet}

\item{topology}{topology to be fixed in MrBayes analysis, specified in any way acceptable to \link{read_tree}.}

\item{interleaved}{logical, whether to write sequence data in interleaved format}

\item{beagle}{logical, whether to use beagle for parallelization, default is FALSE (disable it)}
}
\description{
Writes input file for MrBayes analysis including data matrix and the block of MrBayes commands.
}
\details{
The \code{mcmc} defaults are: ngen=2500000, samplefreq=1000, relburnin="yes", burninfrac=0.1, savebrlens="yes", nchains=4, nruns=4.
The \code{sumt} defaults are: contype="allcompat", minpartfreq="0.0", calctreeprobs="no", showtreeprobs="no", hpd="yes", relburnin="yes",
burninfrac=0.1.
}
