% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seq.R
\name{write_nexus}
\alias{write_nexus}
\title{Export into nexus file.}
\usage{
write_nexus(
  loci,
  file,
  nloci = -1L,
  part = NULL,
  rm.empty = FALSE,
  interleaved = TRUE,
  toupper = TRUE,
  snps = TRUE,
  missing = "N",
  gap = "-"
)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{file}{character string, the name of the .alleles file.}

\item{nloci}{the number of loci to be exported (the default \code{-1L} means all of them)}

\item{part}{either a two column matrix indicating starts and ends of loci in the concatenated aligments
(in .fasta or .nexus files) or a name of partition file formatted according to RAxML or MrBayes convention.}

\item{rm.empty}{logical, whether to remove empty (e.g. all "-" or all "N") sequences.}

\item{interleaved}{not relevant, just for compatibility with \link{import_seq}.}

\item{toupper}{logical, whether to represent nucleotides by upper case letters.}

\item{snps}{logical, whether to (extract and) include information about SNPs.}

\item{missing}{a symbol for missing data.}

\item{gap}{a symbol for gap in sequence.}
}
\description{
Exporting sequence data into nexus file.
}
