% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_part.R
\name{write_part}
\alias{write_part}
\title{Write partition file.}
\usage{
write_part(part, file, format = c("raxml", "nexus"), type = "DNA")
}
\arguments{
\item{part}{a data frame with infomration about partition scheme, formatted as an output of \link{read_part}.}

\item{file}{a name of the partition file, if missing, the output is printed to console.}

\item{format}{a character string, either \code{"raxml"} or \code{"nexus"}.}

\item{type}{currently, only \code{"DNA"} is assumed here.}
}
\description{
Writes nexus or RAxML-formatted partition files.
}
