% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_seq.R
\name{write_seq}
\alias{write_seq}
\title{Exporting sequence data.}
\usage{
write_seq(
  loci,
  file,
  format,
  nloci = -1L,
  part = NULL,
  rm.empty = FALSE,
  interleaved = TRUE,
  toupper = TRUE,
  snps = TRUE,
  missing = "N",
  gap = "-"
)
}
\arguments{
\item{loci}{a list of locus-specific sequence alignments of class \code{"matrix"} or \code{"DNAbin"}.}

\item{file}{character string, the name(s) of the file(s).}

\item{format}{character, the format of the file(s); it can be \code{"ipyrad"}, \code{"nexus"}, \code{"fasta"} or \code{"phylip"}.}

\item{nloci}{the number of loci to be exported (the default \code{-1L} means all of them)}

\item{part}{either a two column matrix indicating starts and ends of loci in the concatenated aligments
(in .fasta or .nexus files) or a name of partition file formatted according to RAxML or MrBayes convention.}

\item{rm.empty}{logical, whether to remove empty (e.g. all "-" or all "N") sequences.}

\item{interleaved}{Logical, whether in interleaved format; relevant for \link{write_nexus} and \link{write_phylip}.}

\item{toupper}{logical, whether to represent nucleotides by upper case letters.}

\item{snps}{logical, whether to (extract and) include information about SNPs.}

\item{missing}{a symbol for missing data.}

\item{gap}{a symbol for gap in sequence.}
}
\description{
The functions exporting sequence data into various file formats.

\code{write_seq} is a wrapper for the following functions.

\code{write_nexus} into .nexus file, with loci defined by an appropriate nexus block.

\code{write_fasta} into .fasta file, loci concatenated or exported into different files.

\code{write_phylip} into .phy file, loci concatenated or exported into different files.

\code{write_ipyrad} into .alleles file used by ipyrad RADseq assembler.
}
\details{
a known issue is a correct specification of loci placed on negative strands in \link{write_ipyrad}
applied to data from \link{import_stacks}. Currently, the functions do not distnguish missing data from gaps,
although they use an appropriate coding convention (\code{missing="N"} and \code{gap="-"}). What is literally missing
in the input is replaced by \code{"N"}, but dashes in input are left as they are, as the accepted inputs do not
contain any information about their actual meaning.
}
