% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_structure.R
\name{write_str_data}
\alias{write_str_data}
\title{Writing STRUCTURE data file.}
\usage{
write_str_data(
  loci,
  file = NULL,
  popdata = NULL,
  popflag = 1,
  locdata = NULL,
  recessive = FALSE,
  linkage = FALSE,
  haploid = NULL,
  allele = "_[012]$",
  return = FALSE
)
}
\arguments{
\item{loci}{input data, a matrix or a list of multiple sequence alignments.}

\item{file}{character string, name of the output file.}

\item{popdata}{a data frame or a matrix with classification of individuals (1st column) into populations (2nd column).}

\item{popflag}{logical, whether to use the popdata, can be a vector of the same length as the number of individuals.}

\item{locdata}{a data frame or a matrix with classification of individuals (1st column) into a priori clusters (2nd column).}

\item{recessive}{logical, whether to specify recessive alleles, or integer, giving the recessive alleles for each locus.}

\item{linkage}{logical, whether to apply linkage model, or numeric, giving distances between successive loci.
If distances are specified, the first locus in each linkage group has value -1.}

\item{haploid}{a list of two vectors containing individual names (1st vector) and loci (2nd vector) with haploid genotypes.}

\item{allele}{regular expression with allele identifier.}

\item{return}{logical, whether to return genotype matrix}
}
\value{
A matrix with genotypes and linkage information coded as in STRUCTURE input file.
}
\description{
Writes STRUCTURE data file.
}
