% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_structure.R
\name{write_str_params}
\alias{write_str_params}
\title{Writing STRUCTURE parameter files.}
\usage{
write_str_params(df, outputs = c("main.txt", "extra.txt"))
}
\arguments{
\item{df}{a data frame with argument names (1st column) and values (2nd column).}

\item{outputs}{character vector, names of the parameter files.}
}
\description{
Prepares STRUCTURE parameter files by accommodating their templates.
}
