% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tree.R
\name{write_tree}
\alias{write_tree}
\title{Write trees.}
\usage{
write_tree(phy, file, format = "newick", translate = TRUE, tree.names = TRUE)
}
\arguments{
\item{phy}{an object of class \code{phylo} or \code{multiPhylo}.}

\item{file}{a character string with the file name.}

\item{format}{character string, either \code{"newick"} (default) or \code{"nexus"}}

\item{translate}{logical, relevant for the nexus format, whether to use numeric representation of tip labels}

\item{tree.names}{logical, relevant for the nexus format, whether to keep tree names}
}
\value{
If \code{file} is not supplied, it returns character vector with newick text strings.
}
\description{
Writes phylogenetic trees into a file written in newick or nexus format or into a newick text string.
}
