% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate sequences.}
\usage{
concatenate(loci, toupper = TRUE, as.DNAbin = FALSE, missing = "-")
}
\arguments{
\item{loci}{a list of alignments in \code{matrix} or \code{DNAbin} format.}

\item{toupper}{logical, whether to export sequences in upper case letters.}

\item{as.DNAbin}{logical, whether to return the alignment in \code{DNAbin} format.}

\item{missing}{a symbol representing missing nucleotides.}
}
\value{
A concatenated alignment in \code{matrix} or \code{DNAbin} format with \code{part} attribute defining locus boundaries.
}
\description{
Concatenates multi-locus sequences.
}
