% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytree.R
\name{estimate_brlen}
\alias{estimate_brlen}
\title{Branch lengths.}
\usage{
estimate_brlen(phy, trees, strict = FALSE, fun = "mean", thin = 1)
}
\arguments{
\item{phy}{a \code{phylo} object with the summary tree (branch lengths are omitted, even if present).}

\item{trees}{a posterior sample of rooted trees in a \code{multiPhylo} object or a list of \code{phylo} objects
or an object of class \code{bpp} or a name of file with the posterior sample.}

\item{strict}{logical, whether to estimate branch lengths only from the trees, whose topology is strictly
identical with that of a summary tree.}

\item{fun}{character string, name of the function estimating branch lengths (default is \code{"mean"}).}

\item{thin}{a proportion of mcmc steps to be retained or a subsampling frequency (if \code{thin > 1}).}
}
\value{
An object of class \code{phylo} amended by branch lengths estimated from the provided sample of trees.
}
\description{
Finds a summary tree for a posterior sample from a Bayesian phylogenetic inference.
}
