% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcomp.R
\name{pcomp}
\alias{pcomp}
\title{Principal component analysis.}
\usage{
pcomp(b)
}
\arguments{
\item{b}{matrix of binary coded genotype data, possibly the output of \link{make_binary_snps}.}
}
\value{
A list with components \code{values} and \code{vectors} with results of the spectral decomposition of
internally computed pairwise inter-individual covariance matrix.
}
\description{
Principal component analysis of binary coded genotype data.
}
\details{
The function is an implementation of PCA as described by Patterson et al. 2006 (PLoS Genet 2: e190).
The spectral decomposition is performed by \link[base:eigen]{base::eigen}.
}
