% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascertbias.R
\name{plot_ascertbias}
\alias{plot_ascertbias}
\title{Ascertainment bias plot.}
\usage{
plot_ascertbias(
  var,
  ascert,
  bin = "occupancy",
  cumulative = TRUE,
  xlab,
  ylab,
  main = "",
  cex = 1.5,
  cex.lab = 1.5,
  cex.axis = 1.25,
  return = FALSE,
  device,
  file
)
}
\arguments{
\item{var}{a name of indicator (y-axis) variable; must be a name of component of the list returned by \link{ascertbias},
i.e. one of the following: 'mdist', 'nalleles', 'nsnps', 'npis', 'heterozygosity', 'spoccupancy', 'mintra'
or (possibly) 'occupancy', or their unambiguous abbreviation.}

\item{ascert}{output of \link{ascertbias}.}

\item{bin}{a name of binning (x-axis) variable. Options are the same as in \code{var}, but the default is 'occupancy'.}

\item{cumulative}{logical, whether to plot statistics for increasingly filtered sets of loci
rather than for different occupancy bins. The default is \code{TRUE}.}

\item{xlab, ylab, main}{labels of axes and the whole plot; if not specified, \code{xlab} and \code{ylab} are given
pre-defined values given the specification of \code{bin} (x-axis) and \code{var} (y-axis) variables.}

\item{cex, cex.lab, cex.axis}{size of points, axis labels and tick labels.}

\item{return}{logical, whether to return a list with binned statistics (\code{classes}), their medians (\code{medians}),
break points between bins (\code{breaks}) and the numbers of loci contained in the bins (\code{nloci}).}

\item{device}{either \code{"quartz"}, \code{"x11"} (= \code{"X11"}) or \code{"pdf"}.}

\item{file}{character strigm the name of pdf file if \code{device == "pdf"}.}
}
\description{
Plots results of the ascertainment bias analysis.
}
