% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clades.R
\name{plot_clades}
\alias{plot_clades}
\title{Clades in the tree.}
\usage{
plot_clades(
  phy,
  clades,
  color,
  bg = "black",
  type = "phylogram",
  edge.color = NULL,
  edge.width = 3,
  show.clade.lines = TRUE,
  show.clade.label = TRUE,
  show.tip.label = TRUE,
  show.node.label = FALSE,
  show.edge.label = FALSE,
  clade_labels = NULL,
  tip_labels = NULL,
  node_labels = NULL,
  edge_labels = NULL,
  root_label = NULL,
  ancestral = TRUE,
  clade.order = FALSE,
  show.scale = TRUE,
  device,
  width = 7,
  height = 7,
  file = NULL,
  outer = NULL,
  inner = NULL,
  clade.lin.pos = 0,
  clade.lin.over = 0.4,
  clade.lab.col = NULL,
  clade.lab.cex = 1.25,
  clade.lab.font = 1,
  clade.lab.las = 2,
  tip.lab.pos = 0.005,
  tip.lab.cex = 1,
  tip.lab.font = 3,
  node.lab.pos = 0,
  node.lab.cex = 1,
  node.lab.font = 1,
  node.lab.col = "black",
  node.lab.border = "black",
  node.lab.bg = "white",
  node.lab.lwd = 1,
  node.lab.asp = 1,
  node.lab.adj = c(0.5, 0.5),
  edge.lab.pos = 0.5,
  edge.lab.cex = 1,
  edge.lab.font = 1,
  edge.lab.col = "black",
  edge.lab.border = "black",
  edge.lab.bg = "white",
  edge.lab.lwd = 1,
  edge.lab.asp = 1,
  edge.lab.adj = c(0.5, 0.5),
  scale.cex = 1,
  scale.pos = c(0.5, 0)
)
}
\arguments{
\item{phy}{an object of class \code{phylo}.}

\item{clades}{vector of clade labels (order must correspond to \code{phy$tip.label}) or a data frame (or matrix)
with tip labels (1st column) and clade assignment (2nd column). The classification need not to be comprehensive.}

\item{color}{named vector of colors. The names correspond to either clade labels or tip labels, the latter is
relevant especially if coloration is unlinked from clade assignment. It can be also a matrix or data frame
with labels (1st column) and colors (2nd column).}

\item{bg}{color of background (unclassified) branches.}

\item{type}{character, tree type, either \code{"phylogram"} (default) or \code{"cladogram"}.}

\item{edge.color}{branch colors, vector of the same length as \code{nrow(phy$edge)}, if specified,
it supersedes \code{color} of branches, while clade lines and names are not affected.}

\item{edge.width}{numeric, branch width.}

\item{show.clade.lines}{logical, whether to show clade lines.}

\item{show.clade.label}{logical, whether to show clade names.}

\item{show.tip.label}{logical, whether to display tip labels.}

\item{show.node.label}{logical, whether to display node labels.}

\item{show.edge.label}{logical, whether to display edge labels.}

\item{clade_labels}{a named vector of clade labels, the names are clade names;
useful to modify clade names for display (e.g. include spaces, delete underscores).}

\item{tip_labels}{vector of tip labels, in order corresponding to \code{phy$tip.label}.}

\item{node_labels}{vector of node labels, in order corresponding to node numbers in \code{phy$edge}.}

\item{edge_labels}{vector of edge labels, in order corresponding to \code{phy$edge}.}

\item{root_label}{a character string added as a root label to \code{node_labels}.}

\item{ancestral}{logical, whether to show \code{node_labels} and \code{edge_labels} only for clade ancestors.}

\item{clade.order}{logical (whether to order the clades alphabetically) or character
(clade labels in the required order).}

\item{show.scale}{logical, whether to display scale bar as x-axis.}

\item{device}{either \code{"quartz"}, \code{"x11"} or \code{"pdf"} to indicate the type of graphical device.
If \code{NULL}, the objects are plotted into the current device.}

\item{width}{width of graphical device in inches.}

\item{height}{height of graphical device in inches.}

\item{file}{a name of pdf file if \code{device == "pdf"}.}

\item{outer}{sizes of outer margins in inches, recycled if necessary.}

\item{inner}{sizes of inner margins as proportions of plot dimensions, recycled if necessary.}

\item{clade.lin.pos}{position of clade lines, negative values ~ shift to the left.}

\item{clade.lin.over}{overhang of clade lines as a proportion of single character height.}

\item{clade.lab.col}{color of clade names.}

\item{clade.lab.cex}{size of clade names.}

\item{clade.lab.font}{font of clade names.}

\item{clade.lab.las}{orientation of clades names relative to the axis (default perpendicular).}

\item{tip.lab.pos}{gap between tree tips and tip labels as a proportion of the plot width (without outer margins).}

\item{tip.lab.cex}{size of tip labels.}

\item{tip.lab.font}{font of tip labels.}

\item{node.lab.pos}{node label position as a proportion of the supporting branches.}

\item{node.lab.cex}{size of node labels.}

\item{node.lab.font}{font of node labels.}

\item{node.lab.col}{color of node labels.}

\item{node.lab.border}{color of node label point border.}

\item{node.lab.bg}{color of node label point background.}

\item{node.lab.lwd}{width of node label point border.}

\item{node.lab.asp}{shape of node label, aspect ratio of an ellipse, default=1 means cricle.}

\item{node.lab.adj}{adjustment of text position in node label (\code{adj} argument of \link[graphics:text]{graphics::text}).}

\item{edge.lab.pos}{edge label position as a proportion of the branch from its start (= root side).}

\item{edge.lab.cex}{size of edge label point.}

\item{edge.lab.font}{font of edge labels.}

\item{edge.lab.col}{color of edge labels.}

\item{edge.lab.border}{color of edge label point border.}

\item{edge.lab.bg}{color of edge label point background.}

\item{edge.lab.lwd}{width of edge label point border.}

\item{edge.lab.asp}{shape of edge label, aspect ratio of an ellipse, default=1 means cricle.}

\item{edge.lab.adj}{adjustment of text position in edge label (\code{adj} argument of \link[graphics:text]{graphics::text}).}

\item{scale.cex}{size of scale labels.}
}
\description{
Plots phylogenetic tree while highlighting defined clades.
}
