% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.R
\name{read_str_data}
\alias{read_str_data}
\title{Reading STRUCTURE data file.}
\usage{
read_str_data(
  file,
  row.names = 1,
  recessive = FALSE,
  linkage = FALSE,
  phase = FALSE
)
}
\arguments{
\item{file}{character string, name of the data file.}

\item{row.names}{a vector of row names or an index of the column containing row names.}
}
\value{
A matrix with genotypes, associated with an attribute \code{info} which contains the first columns with metadata.
}
\description{
Reads STRUCTURE input data file.
}
\details{
It supposes haploid or diploid genotypes.
}
