% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_structure.R
\name{write_str_data}
\alias{write_str_data}
\title{Writing STRUCTURE data file.}
\usage{
write_str_data(
  loci,
  file = NULL,
  linkage = FALSE,
  haploid = NULL,
  allele = "_[012]$",
  PopData = NULL,
  PopFlag = 1,
  return = FALSE
)
}
\arguments{
\item{loci}{input data, a matrix or a list of multiple sequence alignments.}

\item{file}{character string, name of the output file.}

\item{linkage}{logical or numeric, whether to apply linkage model, or numeric, giving distances between successive loci.
If distances are specified, the first locus in each linkage group has value -1.}

\item{haploid}{a list of two vectors containing individual names (1st vector) and loci (2nd vector) with haploid genotypes.}

\item{allele}{regular expression with allele identifier.}

\item{PopData}{a data frame or matrix containg classification of individuals (1st column) into populations (2nd column).}

\item{PopFlag}{logical, whether to use the PopData}

\item{return}{logical, whether to return genotype matrix}
}
\value{
A matrix with genotypes and linkage information coded as in STRUCTURE input file.
}
\description{
Writes STRUCTURE data file.
}
